rename apiConstants to serverRoutes
