+++ src/main/typescript/actions/authenticationActions.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import { ReduxAction } from "../lib/redux"
-import { loginURL, logoutURL } from "../constants/apiConstants"
+import { loginURL, logoutURL } from "../constants/serverRoutes"
 import { AuthenticationConstants } from "../constants/authenticationConstants"
 import axios from "axios"
